Allow to fully parse spec files with missing local sources
RPM is unable to fully parse a spec file if some sources are missing.
Only so called non-build parse is possible in that case, but the result
is not the same, it's like the sources weren't even defined.

Overcome that by creating a temporary dummy source for every source
that is missing locally. Mimic file signatures of compressed sources
to make RPM behave exactly the same when processing a dummy source.

Signed-off-by: Nikola Forró <nforro@redhat.com>